requirements.txt
google.cloud.storage>=1.36.2
google.cloud.bigquery>=2.12.0
